// src/types/developer.ts
export interface IDeveloper {
  image: {
    thumbnail: string
    alt: string
    title: string
    url: string
  }
  capital: number
  website: string
  address: string
  reg_num: string
  director: string
  keyId: string
  business_segment: string
  contact_info: string
  display_name: string
  branch: string
  title_th: string
  bank_id: string
  location: {
    bottom: string
    lon: string
    right: string
    top: string
    left: string
    lat: string
  }
  title_en: string
  id: string
  department: string
  email: string
}